separate crosswidget and multiple viewer widget
